+++ eva/src/store/storeRest.js
@@ -1,4 +1,4 @@
-import { mdiNetwork } from '@mdi/js';
+import { mdiNetwork, mdiAlertCircleOutline } from '@mdi/js';
 import Vue from 'vue';
 
 export default {
@@ -22,6 +22,8 @@ export default {
           formData: Object.fromEntries(formData.entries()),
           searchFrom,
         });
+        console.log('Object.fromEntries(formData.entries())', Object.fromEntries(formData.entries()));
+        console.log('searchFrom', searchFrom);
         worker.onmessage = (event) => {
           // console.log('[response] worker: %s', searchFrom.sid/*, event.data*/)
           worker.terminate();
@@ -36,7 +38,15 @@ export default {
             log.forEach(([time, msg]) => restAuth.putLog(`[worker] ${msg}`, time));
           }
           if (error) {
-            console.log('[error]', error);
+            console.log('event', event);
+            this.store.commit('notify/addNotification', {
+              id: searchFrom.sid,
+              icon: mdiAlertCircleOutline,
+              message: error.replaceAll('&nbsp;', ' '),
+              read: false,
+              type: 'error',
+              time: 2,
+            });
             restAuth.putLog(error);
           }
           if (notifications.length) {