+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Table event OnDataCompare: 
   - Ability use the token in the value
   - Ability to use regular expressions in the column name
+- Notify on data source error
 
 ### Changed
 - removed libraries yFiles and license from project